Android: move GamesAdapter class to own file
